Fix copy link command validation
The validation does not account for the current page. For example,
the user may be on page 2 and wish to copy result 17. The logic
treats this as out-of-bounds because there are only ever 10 results.

This change takes the current page into account when calculating the
result index to copy the link from.